Fix throbber.gif path for mediapicker-modal in the webpage environment in webapp
Summary:
The path for the throbber.gif for the webpage environment in webapp contains
an extra "/" which causes it to 404.

Test Plan:
Load http://localhost:8080/computer-programming/new/webpage and add an image
using an image tag and the mediapicker.  Open the console and verify that there
isn't a 404 containing "//throbber.gif"

Reviewers: john

Subscribers: pamela

Differential Revision: https://phabricator.khanacademy.org/D20268