Fix line numbers reported by Program.assertEquals
Summary:
The recent addition of LoopProtector caused the line numbers reported
by Program.assertEquals to be wrong because additional code was being
inserted into user programs that Program.assertEquals was unaware of.
This diff fixes that by passing the line and column number directly
from the AST to calls to Program.assertEquals by adding two arguments
to the AST before it's codegen-ed.  By using the locations for the AST
we remove the need for TraceKit.

https://app.asana.com/0/28126167511849/44955730861534

Test Plan: gulp test_output_pjs

Reviewers: pamela

Reviewed By: pamela

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D20401