Fix call to walkAST in LoopProtector::protect.
Summary:
The pjs environment doesn't use the protect method, but the webpage environment
does and I forgot to run the webpage tests.  This diff updates the call to walkAST
in the protect method so that the right number of args are passed to it.

Test Plan: - gulp test_output_webpage

Reviewers: john, pamela

Reviewed By: john, pamela

Differential Revision: https://phabricator.khanacademy.org/D20436