Remove 'with' statement from user code wrapper.
Summary:
We were using 'with' to inject processing.js methods into the execution
context where we run user code.  This is so that users don't have to write
`processing.fill(...)` and can instead write `fill(...)`.  Unfortunately,
using 'with' has a pretty severe performance impact which was highlight by
the addition of more code by the loop protector.

In order to get performance to an acceptable level we replace the 'with'
statement with auto-prefixing user code with '__env__<RandomInt>'.  This
is done by analyzing and up the AST for the user code.

Test Plan: - gulp test_output_pjs

Reviewers: pamela, john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D20597