Don't prefix 'arguments' with __env__9182309235
Summary:
All globals are prefixed with the context name, e.g. __env__1932879128.
We were incorrectly identifying 'arguments' as a global as prefixing it
with the context name when in fact it's a special object that exists inside
functions which contains all of the arguments passed to the function.
This diff prevents 'arguments' from being prefixed.

Test Plan: - gulp test_output_pjs

Reviewers: john, pamela

Reviewed By: pamela

Differential Revision: https://phabricator.khanacademy.org/D20819