Fix logic bug that was setting this.validation to undefined in live-editor.js
Summary:
The original code used `!=` to check for both `undefined` and `null` simultaneously.
At some point this check was changed to `!== null` which meant that it was accepting
`undefined` instead of rejecting it.  I've changed it to be quite explicit so that
we're not relying on JavaScript's type coercion while maintaining the original
intended behavior.

Test Plan:
- load http://localhost:8080/computing/computer-programming/programming/drawing-basics/p/challenge-h-for-hopper
- go through the challenge and verify that the "Next Step!" button appears for each step

Reviewers: johnsullivan

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D21344