Add an option to enable/disable to LoopProtector so that official programs are allowed to run long.
Summary:
Some of the programs in Pixar in a Box are compute intensive and may
take too long to run on some machines.  In order to prevent the LoopProtector
from stopping the program and throwing up an error message we provide
an option to enable/disable with the default being enabled.  This will be
used by scratchpad-ui.js in webapp to disable the LoopProtector for this
programs that are canvasOnly (future diff).

Test Plan:
- modify dmoes/simple/index.html so that the options to new LiveEditor include 'enableLoopProtect: false'
- open the pjs demo and add an infinite loop and see that it hangs the browser
- remove the line and reload
- see that the LoopProtector kicks in

Reviewers: john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D21306