Fix the jQuery challenges by using the DOM Parser to parse user code
Summary:
We were looking at the DOM within the output iframe after setting the
document via frameDoc.write(userCode).  Unfortunately, some browsers
update the DOM asynchronously and so the changes aren't always guaranteed
to be in the DOM before we tried to grab all of the <script> tags to
get the JavaScript code for validation.

This diff use DOM Parser to parse the user code into an document fragment.
Since this process is synchronous it's much more reliable.

Test Plan:
- gulp test_output_webpage (I added a jQuery test.  Please note that this was not a failing test before the fix mainly because the test harness had a certain amount of asychronousity built into it)
- manual testing:
   - open up http://localhost:8080/computer-programming/challenge-animated-avatar/6741719346446336 in Firefox
   - insert `$("img").animate({width:500});` into the empty <script> tag
   - hit enter a whole bunch of times, then delete, then alternate (spam it!)
   - verify that "next step" Hopper remains on the page throughout

Reviewers: john, pamela

Reviewed By: pamela

Differential Revision: https://phabricator.khanacademy.org/D21709