Extract code injection code from pjs-output.js so that it's more modular
Summary:
We'd like to be able to use official CS scratchpads on mobile.  In order
to make scratchpads load faster, we'd like to be able to load code that's
already been transformed.  In order to transform code ahead of time, it
would be nice to be able to use the code injection code which transforms
the code independently of pjs-output.  Extracting this code into its own
module is the first step in achieving this goal.

The next step (in a following diff) would be to extract the code that does
the AST transform so that it can be called separately and add an option so
that it can generate standalone code which requires some tweaks.  See 1.b and 1.c
in the "Next Steps" section of
https://docs.google.com/document/d/1xEw8Q-WjQa_qGHkj56je08zQ0fP6ljXS5AXh5-BZmVw/edit#heading=h.d21a8p7n3j3o

Test Plan: - gulp test_output_pjs

Reviewers: john, pamela

Reviewed By: pamela

Differential Revision: https://phabricator.khanacademy.org/D22180