Refactor test runners so that individual test suites can be run one at a time
Summary:
In order to improve test reliability when running from the command line and
especially when running on travis-ci we'd like to be able to run fewer tests
at time before using phantomjs.  The theory is that fewer tests will require
less memory to run.  At the end of each suite we restart phantomjs in order
to clear the memory we've accumulated on the heap.

test results: https://travis-ci.org/Khan/live-editor/builds/85978959

Test Plan:
- npm test
- create a pull request to check travis-ci integration

Reviewers: john, pamela

Reviewed By: pamela

Differential Revision: https://phabricator.khanacademy.org/D22509