Update test-worker.js to have i18n._
Summary:
test-worker.js is used to run tests on user code for challenges
and return those results asynchronously.  This diff adds i18n._
in addition to $._ so that test code for challenges can be transitioned
over to use i18n._.  Once that work is complete, the $._ function
should be deleted.

Test Plan:
- load http://localhost:8080/computing/computer-programming/programming/variables/p/challenge-funky-frog?lang=pl
- verify that challenge test strings show up translated (I've modified this one challenge to use i18n._ instead of $._)

Auditors: pamela csilvers john