Cause travis-ci tests to fail when there is one or more test failures
Summary:
This diff communicates how many test failures occur in each suite
by throwing an error in the gulp-mocah-runner.js plugin.  This error
is caught and the number of failures is recorded.  All suites are
run and the number of failures is returned as an error code at the
end.  There are 2 tests that are failing.  I will fix these tests in
a followup diff.

Test Plan:
- run `npm test` and verify the return code is 1 by running `echo $?`
- push so that the tests run on travis, verify that they fail

Reviewers: pamela john