Fix failing tests now that travis-ci fails the build when tests fail
Summary:
Two of the output tests had been failing.  The defaultTimeout was
ignoring the loopProtectTimeout that was being passed in by the tests.
Since the defaultTimeout's initialTimeout value was 2000ms, this was
causing the test's which also have a 2000ms timeout to fail.  This
diff removes the defaultTimeout and uses the loopProtectTimeout option
passed to PJSCodeInjector's constructor instead.

Test Plan:
- npm test
- set a breakpoint in PJSCodeInjector's constructor and load demos/simple/index.html and verify that this.loopProtectTimeout is `{ initialTimeout: 2000, frameTimeout: 500 }`

Reviewers: john, pamela

Reviewed By: pamela

Differential Revision: https://phabricator.khanacademy.org/D23164