Fix failing test exporting test
Summary:
A recent change introduced an AST transform for converting `new`
expression to `PJSOutput.applyInstance` calls.  This change relies
on `PJSOutput` being defined on the context.  The failing test
was not adding PJSOutput to the context b/c it use to work fine
as a global.  The fix is to not do the transform in the first
place for code that's being exported b/c it's not necessary.  In
fact, it's actually preferrable not to do this transform because
the code isn't idiomatic JavaScript and will probably confuse
users who export their code in the future.

Test Plan: - gulp test_output_pjs

Reviewers: pamela, john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D23263