Resolve regression bug #496
https://github.com/Khan/live-editor/issues/496 was caused due to the
fact that the AST transform created in
https://github.com/Khan/live-editor/commit/cd8f6dc529f387ad107a610a947254ff5ddbdbb2,
was always prefixing identifiers with '__env__'.
Function parameters are not prefixed with '__env__', therefore trying to
reference '__env__.someFunctionParam' would inevitably fail.

This commit adds a ternary-if-statement that checks to see if we're
inside of a 'FunctionExpression'. If we are, we don't prefix identifiers
when we rewrite 'NewExpression's to 'CallExpression's.