Fix a few issues from recent OhNoes reorg
Summary:
I was noticing some oddness after the new OhNoes, and I think I figured it out:
1) Now that we call maybeShowErrors from another place, there's a possibility
that the erorr timeout doesn't get cleared. So I always clear it before setting it now,
plus I added a check in there to not go into errorstate unless we have errors,
just to be on the safe side.
2) I also added a conditional check based on an error I saw in Sentry and couldn't replicate -
"undefined is not an object (evaluating 'this.editor.getCursor().row'`)*​"
  It happened in the blocks challenges, so it could be a case of no editor or no cursor object.

Test Plan:
- Type ellipse(, click OhNoes, see error pop up, keep typing, see OhNoes go down.
- Watch More Drawing talk-through, never see oh noes disappear at bottom.

Reviewers: kevinb

Subscribers: #cs

Differential Revision: https://phabricator.khanacademy.org/D23900