Protect against more potential XSS in error-buddy loop check
Summary:
While D23874 made it a lot harder to make error buddy execute arbitrary code
outside the iframe, it didn't prevent it entirely.  In particular, we were
still doing the sanitization inside the iframe, which means that we would have
to be really careful when sanitizing the object (since it could have arbitrary
getters, toString attributes, non-writeable properties, and so on), to prevent
code like the following:
```
throw Object.defineProperty({}, "html", {
    configurable: false,
    value: "<h1>oh noes!</h1><script>console.log(location.href);</script>"
});
```
But in addition to that, a user could just pass their own message up to the
parent and avoid our sanitization entirely:
```
(function() {
    // we need to wait until after the "program ended" message gets sent, or
    // our message will be quickly removed
    this.setTimeout(function() {
        this.parent.postMessage(this.JSON.stringify({
            results: {
                errors: [{
                    text: "<h1>oh noes!</h1><script>console.log(location.href);</script>",
                    row: 8
                }]
            }
        }), "*");
    }, 1000);
})();
```
Now we'll do the sanitization outside the iframe.  We'll still need to be
careful to sanitize any strings we display to the user and to do some
sanity-checking to make sure that we aren't, say, assuming that something is
safe to string-format without further sanitization since we expect it to be a
number (when in fact the user could pass us a string).  But since the messages
are JSON-formatted, we can assume that everything in sight is a builtin plain
object.

I'd still like to be a bit less ad-hoc about our sanitization, and consolidate
it to make it easier to reason about and to check that everything is sanitized;
I left a TODO for that.  I also left a TODO to write some tests to assert that
we don't display HTML that comes from the iframe.

I also threw what I think is a fix for #518/#522 which seem to be regressions
from D23874, but I can't seem to get consistent repro of either so I'm not
entirely sure if it solves the problems.

Test Plan:
Substate into webapp, then:
* Run the above code snippets, make sure nothing gets logged to the console.
* Write an infinite loop, throw an error, and make a typo, and make sure error
  buddy still works normally in all cases (both in pjs and webpage).
* `gulp test`
* Load `tests/output/pjs/index.html?tests=async_test.js` in a browser, check
  that the only failures are expected ones.

Reviewers: jlfwong, pamela, kevinb

Reviewed By: kevinb

Subscribers: sean

Differential Revision: https://phabricator.khanacademy.org/D24252