Fix #504, tranform of NewExpressions where the contructor is a MemberExpression fails
Summary:
We were assuming that the callee for the NewExpression was always a string.
This was not a safe assumption to make because it can be a MemberExpression.
This diff handles that case by converting the MemberExpression to a string
before passing it to PJSOutput.applyInstance.  I've also added a test case.

Auditors: pamela