Add some missing avatars to our image picker
Running a comparison of `all-images.js` and `images.js`, we appear to
be missing some avatars from our picker. This adds "avatar-team" and
"mr-pants-with-hat".

I've left "mystery-1" and "mystery-2" out because they sound like
they're meant to be a secret. Also, the images don't seem as fun as
the others : )

Test Plan:

- Run `python -m SimpleHTTPServer`
- Go to http://localhost:8000/demos/simple/
- Type `image(getImage(""), 0, 100, 400, 200);`
- Click to see the image picker

"avatar-team" and "mr-pants-with-hat" should appear.

Auditors: dhruv