Change out throbber for spinner
Summary: We are updating the loading animation we use on KA and would like to update the live editor throbber as well. Since the live editor only uses gif versions of the animations, we just need to swap in the new gifs.

Test Plan: load up http://0.0.0.0:1111/demos/simple/ and confirm that we see the new spinner animation

Reviewers: john, kevinb

Subscribers: lauren, tom

Differential Revision: https://phabricator.khanacademy.org/D35040