Set the height of the SVG icons on the scrubber
Summary:
IE has a default height for SVG elements and will use that height if
none is specified.  We were only specifying the width of these elements
b/c that was enough to ensure they were the right size in other browsers.
This diff adds the height as well so things render correctly in IE as well.

BEFORE
{F735891}

AFTER
{F735892}

Test Plan:
- gulp watch
- load http://localhost:8080/computer-programming/new/pjs in IE
- add some code
- use the scrubber on one of the numbers, see that it's the right size

Reviewers: charlie, joshua

Reviewed By: charlie

Differential Revision: https://phabricator.khanacademy.org/D36661