On dev, make sure that postParent really posts to the parent window.
Summary:
On prod, the live editor lives in kasandbox.org on a different domain.
When postFrame gets called in webapp code, the source of the event is
the window at the time or www.khanacademy.org.

On dev however the postFrame code, while it should be executed from
within the parent window, somehow is being executed within the iframe.
It seems React from the very beginning of the stack frame is executing
within the iframe.  It seems like that might be React's fault.

At any rate I put in a temporary hack here until that can be figured
out.  If we notice that the frameSource is actually the same window, then we
must not really have the parent, and can handle that differently.

Test Plan:
http://localhost:8080/computing/computer-programming/programming/drawing-basics/p/drawing-more-shapes-with-code
Hit spin off.  See the title change and the page redirect.

Reviewers: john

Subscribers: #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D38942