Add pointer-events:none; to hidden font container in processing-js
Summary:
opacity:0; isn't enough for Firefox to ignore processing-js' hiddent font
container which results in the text cursor showing up and mouse
events not getting through to the canvas. This diff adds
pointer-events:none; to the container so that Firefox ignores
it completely.

Change to processing-js is in https://github.com/Khan/processing-js/pull/11.

Test Plan:
- npm run build
- load http://localhost:8080/computer-programming/new/pjs (in Firefox)
- hover over the upper-left corner of the canvas
- see that the mouse cursor is always the arrow

Reviewers: joshua, idr

Reviewed By: idr

Differential Revision: https://phabricator.khanacademy.org/D41143