Uses requestAnimationFrame to prevent frozen scripts while seeking (#665)
* Uses window.requestAnimationFrame to prevent frozen script errors while seeking

We've had some reports about the screen locking up while seeking ahead. This is caused by processing commands and building the cache, which trap us in the single thread and prevent other behaviour.

This approach uses `window.requestAnimationFrame` to process a few commands and cache entries, then see if we need to update the browser before trying another pass.

I played with the value of `animationStep` and this seemed okay! It might make sense to show more or less though, so that would be a value to play around with. There's also some place where it's possible to generate bad cache entries if you continue to seek while it's still processing stuff. You can see this if you visit a Khan page like http://localhost:8081/computing/computer-programming/programming/drawing-basics/p/making-drawings-with-code, seek ahead to the end, and while it's processing, start scrolling around the scrollbar. That might be a problem in our Khan `webapp` though; I wasn't sure if I missed returning out of somewhere here in `live-editor` while seeking like in the `play` function, or if this is coming from the scroll bar in `webapp`.

* Remove extra comment, add descriptive comment for new code