Disable same-source postMessage inside output (#706)
Test Plan:
It's really hard to test this in dev, it requires different frame origins. Will attempt to test this in production. At least with this change it should be possible to load a CS page and not see an error in the console.

Additionally, CS challenges should continue to work (they use a lot of postMessages so it's a good confirmation).