Remove global i18n references and support Node v12
We had a couple references to a global i18n object, but this is easy enough to fix!

Also, in order to get the build running, I had to pin the version of graceful-fs to v4 and use npm-shrinkwrap.json to force the sub-dependencies to use it too.

Issue: none

Test plan:
Hopefully this should "just work" inside Webapp?