[🔥AUDIT🔥] Wait for output frame to be active before beginning execution. (#756)
🖍 _This is an audit!_ 🖍

## Summary:
We were waiting for the frame to load, which was happening too early with our recent move to React. We should wait for the frame to tell us we're active.

Issue: FEI-4639

## Test plan:
I imported this into Webapp and loaded up:
http://localhost:8088/computing/computer-programming/html-css-js/html-js-dom-modification/pc/challenge-avatar-attributes

I added in some logging and confirmed that we now wait for the "active" event to come back before we send off code to be executed. In my limited testing so far this seems to be very successful!

Author: jeresig

Auditors: kevinbarabash

Required Reviewers:

Approved By:

Checks:

Pull Request URL: https://github.com/Khan/live-editor/pull/756